# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Amrutam Telemedicine Backend","description":"Production-grade backend for Amrutam telemedicine workflows","version":"1.0.0"},"paths":{"/":{"get":{"tags":["Health"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["Health"],"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/ready":{"get":{"tags":["Health"],"summary":"Readiness Check","operationId":"readiness_check_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health/redis":{"get":{"tags":["Health"],"summary":"Redis Health Check","operationId":"redis_health_check_health_redis_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/register":{"post":{"tags":["Auth"],"summary":"Register User","operationId":"register_user_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegisterRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Auth"],"summary":"Login User","operationId":"login_user_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["Auth"],"summary":"Get Me","operationId":"get_me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/doctors":{"get":{"tags":["Doctors"],"summary":"Search Doctors","operationId":"search_doctors_doctors_get","parameters":[{"name":"specialization","in":"query","required":false,"schema":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Specialization"}},{"name":"min_rating","in":"query","required":false,"schema":{"anyOf":[{"type":"number","maximum":5,"minimum":0},{"type":"null"}],"title":"Min Rating"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DoctorResponse"},"title":"Response Search Doctors Doctors Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/doctors/availability":{"post":{"tags":["Doctors"],"summary":"Create Availability","operationId":"create_availability_doctors_availability_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvailabilityCreateRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvailabilityResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/doctors/{doctor_id}/slots":{"get":{"tags":["Doctors"],"summary":"Get Doctor Slots","operationId":"get_doctor_slots_doctors__doctor_id__slots_get","parameters":[{"name":"doctor_id","in":"path","required":true,"schema":{"type":"integer","title":"Doctor Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AvailabilityResponse"},"title":"Response Get Doctor Slots Doctors  Doctor Id  Slots Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/consultations/book":{"post":{"tags":["Consultations"],"summary":"Book Consultation","operationId":"book_consultation_consultations_book_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","title":"Idempotency-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/consultations":{"get":{"tags":["Consultations"],"summary":"List Consultations","operationId":"list_consultations_consultations_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"status_filter","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/ConsultationStatus"},{"type":"null"}],"title":"Status Filter"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ConsultationResponse"},"title":"Response List Consultations Consultations Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/consultations/{consultation_id}":{"get":{"tags":["Consultations"],"summary":"Get Consultation","operationId":"get_consultation_consultations__consultation_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"consultation_id","in":"path","required":true,"schema":{"type":"integer","title":"Consultation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/consultations/{consultation_id}/start":{"patch":{"tags":["Consultations"],"summary":"Start Consultation","operationId":"start_consultation_consultations__consultation_id__start_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"consultation_id","in":"path","required":true,"schema":{"type":"integer","title":"Consultation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/consultations/{consultation_id}/complete":{"patch":{"tags":["Consultations"],"summary":"Complete Consultation","operationId":"complete_consultation_consultations__consultation_id__complete_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"consultation_id","in":"path","required":true,"schema":{"type":"integer","title":"Consultation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/consultations/{consultation_id}/cancel":{"patch":{"tags":["Consultations"],"summary":"Cancel Consultation","operationId":"cancel_consultation_consultations__consultation_id__cancel_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"consultation_id","in":"path","required":true,"schema":{"type":"integer","title":"Consultation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/consultations/{consultation_id}/prescription":{"post":{"tags":["Consultations"],"summary":"Create Prescription","operationId":"create_prescription_consultations__consultation_id__prescription_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"consultation_id","in":"path","required":true,"schema":{"type":"integer","title":"Consultation Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrescriptionCreateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrescriptionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Consultations"],"summary":"Get Prescription","operationId":"get_prescription_consultations__consultation_id__prescription_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"consultation_id","in":"path","required":true,"schema":{"type":"integer","title":"Consultation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrescriptionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/payments/{payment_id}":{"get":{"tags":["Payments"],"summary":"Get Payment","operationId":"get_payment_payments__payment_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"payment_id","in":"path","required":true,"schema":{"type":"integer","title":"Payment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/payments/mock-confirm":{"post":{"tags":["Payments"],"summary":"Mock Confirm Payment","operationId":"mock_confirm_payment_payments_mock_confirm_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MockPaymentConfirmRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/admin/analytics/summary":{"get":{"tags":["Admin"],"summary":"Analytics Summary","operationId":"analytics_summary_admin_analytics_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminAnalyticsResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/admin/audit-logs":{"get":{"tags":["Admin"],"summary":"Get Audit Logs","operationId":"get_audit_logs_admin_audit_logs_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/metrics":{"get":{"summary":"Metrics","description":"Endpoint that serves Prometheus metrics.","operationId":"metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AdminAnalyticsResponse":{"properties":{"total_users":{"type":"integer","title":"Total Users"},"total_doctors":{"type":"integer","title":"Total Doctors"},"total_consultations":{"type":"integer","title":"Total Consultations"},"total_prescriptions":{"type":"integer","title":"Total Prescriptions"},"total_payments":{"type":"integer","title":"Total Payments"}},"type":"object","required":["total_users","total_doctors","total_consultations","total_prescriptions","total_payments"],"title":"AdminAnalyticsResponse"},"AvailabilityCreateRequest":{"properties":{"start_time":{"type":"string","format":"date-time","title":"Start Time"},"end_time":{"type":"string","format":"date-time","title":"End Time"}},"type":"object","required":["start_time","end_time"],"title":"AvailabilityCreateRequest"},"AvailabilityResponse":{"properties":{"id":{"type":"integer","title":"Id"},"doctor_id":{"type":"integer","title":"Doctor Id"},"start_time":{"type":"string","format":"date-time","title":"Start Time"},"end_time":{"type":"string","format":"date-time","title":"End Time"},"status":{"$ref":"#/components/schemas/SlotStatus"}},"type":"object","required":["id","doctor_id","start_time","end_time","status"],"title":"AvailabilityResponse"},"BookingRequest":{"properties":{"slot_id":{"type":"integer","exclusiveMinimum":0.0,"title":"Slot Id"},"reason":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Reason"}},"type":"object","required":["slot_id"],"title":"BookingRequest"},"ConsultationResponse":{"properties":{"id":{"type":"integer","title":"Id"},"patient_id":{"type":"integer","title":"Patient Id"},"doctor_id":{"type":"integer","title":"Doctor Id"},"slot_id":{"type":"integer","title":"Slot Id"},"status":{"$ref":"#/components/schemas/ConsultationStatus"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason"}},"type":"object","required":["id","patient_id","doctor_id","slot_id","status","reason"],"title":"ConsultationResponse"},"ConsultationStatus":{"type":"string","enum":["BOOKED","ONGOING","COMPLETED","CANCELLED"],"title":"ConsultationStatus"},"DoctorResponse":{"properties":{"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"specialization":{"type":"string","title":"Specialization"},"experience_years":{"type":"integer","title":"Experience Years"},"consultation_fee":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Consultation Fee"},"rating":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Rating"},"is_verified":{"type":"boolean","title":"Is Verified"}},"type":"object","required":["id","user_id","specialization","experience_years","consultation_fee","rating","is_verified"],"title":"DoctorResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","maxLength":72,"minLength":8,"title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"MockPaymentConfirmRequest":{"properties":{"payment_id":{"type":"integer","exclusiveMinimum":0.0,"title":"Payment Id"},"status":{"$ref":"#/components/schemas/PaymentStatus","default":"PAID"}},"type":"object","required":["payment_id"],"title":"MockPaymentConfirmRequest"},"PaymentResponse":{"properties":{"id":{"type":"integer","title":"Id"},"consultation_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Consultation Id"},"patient_id":{"type":"integer","title":"Patient Id"},"amount":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Amount"},"status":{"$ref":"#/components/schemas/PaymentStatus"}},"type":"object","required":["id","consultation_id","patient_id","amount","status"],"title":"PaymentResponse"},"PaymentStatus":{"type":"string","enum":["PENDING","PAID","FAILED","REFUNDED"],"title":"PaymentStatus"},"PrescriptionCreateRequest":{"properties":{"medicines":{"items":{"additionalProperties":true,"type":"object"},"type":"array","maxItems":20,"minItems":1,"title":"Medicines"},"notes":{"anyOf":[{"type":"string","maxLength":2000},{"type":"null"}],"title":"Notes"}},"type":"object","required":["medicines"],"title":"PrescriptionCreateRequest"},"PrescriptionResponse":{"properties":{"id":{"type":"integer","title":"Id"},"consultation_id":{"type":"integer","title":"Consultation Id"},"doctor_id":{"type":"integer","title":"Doctor Id"},"patient_id":{"type":"integer","title":"Patient Id"},"medicines":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Medicines"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["id","consultation_id","doctor_id","patient_id","medicines","notes"],"title":"PrescriptionResponse"},"SlotStatus":{"type":"string","enum":["AVAILABLE","BOOKED","CANCELLED"],"title":"SlotStatus"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"UserRegisterRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"password":{"type":"string","maxLength":72,"minLength":8,"title":"Password"},"full_name":{"type":"string","maxLength":255,"minLength":2,"title":"Full Name"},"role":{"$ref":"#/components/schemas/UserRole","default":"PATIENT"},"specialization":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Specialization"},"experience_years":{"anyOf":[{"type":"integer","maximum":60.0,"minimum":0.0},{"type":"null"}],"title":"Experience Years","default":0},"consultation_fee":{"anyOf":[{"type":"number","maximum":100000.0,"minimum":0.0},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Consultation Fee","default":0}},"type":"object","required":["email","password","full_name"],"title":"UserRegisterRequest"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","format":"email","title":"Email"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"role":{"$ref":"#/components/schemas/UserRole"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["id","email","phone","role","is_active"],"title":"UserResponse"},"UserRole":{"type":"string","enum":["PATIENT","DOCTOR","ADMIN"],"title":"UserRole"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}},"tags":[{"name":"Auth","description":"User registration, login, and current user APIs"},{"name":"Doctors","description":"Doctor search and availability management"},{"name":"Consultations","description":"Booking, consultation lifecycle, and prescriptions"},{"name":"Payments","description":"Mock payment workflow"},{"name":"Admin","description":"Admin analytics and audit logs"},{"name":"Health","description":"Health, readiness, Redis, and metrics checks"}]}